danis

+++ database/migrations/2024_12_27_051206_update_services_table_for_categories.php
@@ -9,14 +9,28 @@ class UpdateServicesTableForCategories extends Migration
     public function up()
     {
         Schema::table('services', function (Blueprint $table) {
-            $table->json('categories')->nullable()->after('list'); // Добавляем колонку для категорий
+            // Удаляем старое поле 'categories', если оно уже существует
+            if (Schema::hasColumn('services', 'categories')) {
+                $table->dropColumn('categories');
+            }
+
+            // Добавляем три JSON-поля для категорий по языкам
+            $table->json('categories_ru')->nullable()->after('list');
+            $table->json('categories_en')->nullable()->after('categories_ru');
+            $table->json('categories_tm')->nullable()->after('categories_en');
         });
     }
 
     public function down()
     {
         Schema::table('services', function (Blueprint $table) {
-            $table->dropColumn('categories'); // Откат изменений
+            // Удаляем новые поля
+            $table->dropColumn('categories_ru');
+            $table->dropColumn('categories_en');
+            $table->dropColumn('categories_tm');
+
+            // Восстанавливаем старое поле 'categories', если оно было
+            $table->json('categories')->nullable()->after('list');
         });
     }
 }
